revert: load app-picker list synchronously instead of async
PackageManager.getPackages() is fast enough in practice that moving it off
the main thread (a chain of 5 commits reacting to automated review feedback)
bought no measurable benefit on either a slow or fast test device, while
introducing a real search-filter race, a hand-rolled generation counter
reinventing Job cancellation, and weakened test coverage of the guard it
added. Reverted AppPickerActivity.setAppPickerType to the pre-async
synchronous behavior; kept the unrelated action-icon Drawable fix.

Also adds a shared TestDispatchersModule (@TestInstallIn) so any
@HiltAndroidTest gets a deterministic dispatcher automatically instead of
each test class declaring its own local override, and keeps
AppPickerActivityTest/LateralActivitiesTest/LateralActivitiesScreenshotTest
on @HiltAndroidTest as the correct default for testing @AndroidEntryPoint
components, independent of the now-removed dispatcher question.
