+++ app/src/test/java/de/lemke/oneuisample/TestDispatchersModule.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2022-2026 Leonard Lemke
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.lemke.oneuisample
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.components.SingletonComponent
+import dagger.hilt.testing.TestInstallIn
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+
+/**
+ * Replaces [DispatchersModule] in `@HiltAndroidTest` Robolectric tests so coroutines launched on
+ * [IoDispatcher]/[DefaultDispatcher] run unconfined instead of on real background threads -
+ * required so `shadowOf(Looper.getMainLooper()).idle()` can observe their result deterministically.
+ */
+@OptIn(ExperimentalCoroutinesApi::class)
+@Module
+@TestInstallIn(components = [SingletonComponent::class], replaces = [DispatchersModule::class])
+object TestDispatchersModule {
+    @Provides
+    @IoDispatcher
+    fun provideIoDispatcher(): CoroutineDispatcher = UnconfinedTestDispatcher()
+
+    @Provides
+    @DefaultDispatcher
+    fun provideDefaultDispatcher(): CoroutineDispatcher = UnconfinedTestDispatcher()
+}
