+++ app/src/test/java/de/lemke/oneuisample/ui/LateralActivitiesTest.kt
@@ -20,11 +20,14 @@ import android.os.Looper
 import android.view.MenuItem
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
-import de.lemke.oneuisample.App
+import dagger.hilt.android.testing.HiltAndroidRule
+import dagger.hilt.android.testing.HiltAndroidTest
+import dagger.hilt.android.testing.HiltTestApplication
 import de.lemke.oneuisample.R
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
+import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
@@ -32,11 +35,15 @@ import org.robolectric.Shadows.shadowOf
 import org.robolectric.annotation.Config
 import org.robolectric.annotation.GraphicsMode
 
+@HiltAndroidTest
 @RunWith(RobolectricTestRunner::class)
-@Config(application = App::class, sdk = [36])
+@Config(application = HiltTestApplication::class, sdk = [36])
 @GraphicsMode(GraphicsMode.Mode.NATIVE)
 class LateralActivitiesTest {
-    private val context get() = ApplicationProvider.getApplicationContext<android.app.Application>()
+    @get:Rule(order = 0)
+    val hiltRule = HiltAndroidRule(this)
+
+    private val context get() = ApplicationProvider.getApplicationContext<HiltTestApplication>()
 
     private inline fun <reified T : android.app.Activity> launch() {
         ActivityScenario.launch<T>(Intent(context, T::class.java)).use {
